refactor(uibackend): modify code to use new paths
