+++ uibackend/server/cmd/main.go
@@ -34,9 +34,8 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/openclarity/vmclarity/pkg/shared/backendclient"
-	"github.com/openclarity/vmclarity/pkg/uibackend"
-	"github.com/openclarity/vmclarity/pkg/uibackend/api/server"
-	"github.com/openclarity/vmclarity/pkg/uibackend/rest"
+	"github.com/openclarity/vmclarity/uibackend"
+	"github.com/openclarity/vmclarity/uibackend/server/pkg"
 	"github.com/openclarity/vmclarity/utils/log"
 	"github.com/openclarity/vmclarity/utils/version"
 )
